+++ cli/docs/changelog.md
@@ -1,5 +1,8 @@
 # TBD
 
+### Breaking changes
+* Upgrade to core to 1.57.2 and engine to 1.30.1 and container-engine-lib to 0.35.0
+  * Users will need to restart their Kurtosis engine
 # 0.23.2
 ### Fixes
 - Fix a bug when the CLI checks if it's the latest version