+++ cli/docs/changelog.md
@@ -1,7 +1,7 @@
 # TBD
 
 ### Breaking changes
-* Upgrade to core to 1.57.4 and engine to 1.31.1 and container-engine-lib to 0.36.0
+* Upgrade to core to 1.57.5 and engine to 1.31.2 and container-engine-lib to 0.36.0
   * Users will need to restart their Kurtosis engine
 
 # 0.23.2