+++ cli/docs/changelog.md
@@ -22,7 +22,7 @@
 ### Breaking Changes
 * Upgrade to core to 1.57.0 and engine to 1.29.0
   * Users will need to restart their Kurtosis engine
-  
+   
 ### Feature
 * Add test for setting resource allocations when adding a service to an enclave to internal testsuite