+++ cli/docs/changelog.md
@@ -1,5 +1,7 @@
 # TBD
 
+# 0.24.0
+
 ### Breaking changes
 * Upgrade to core to 1.57.6 and engine to 1.31.3 and container-engine-lib to 0.36.1
   * Users will need to restart their Kurtosis engine